POST methods are required for toggling and deleting key.

+++ passkeys/views.py
@@ -13,16 +13,21 @@ def index(request,enroll=False): # noqa
 @require_http_methods(["POST"])
 @login_required
 def delKey(request):
-    key=UserPasskey.objects.get(id=request.GET["id"])
+    id=request.POST.get("id")
+    if not id:
+        return HttpResponse("Error: You are missing a key", status=403)
+    key=UserPasskey.objects.get(id=id)
     if key.user.pk == request.user.pk:
         key.delete()
         return HttpResponse("Deleted Successfully")
     return HttpResponse("Error: You own this token so you can't delete it", status=403)
 
-# @require_http_methods(["POST"])
+@require_http_methods(["POST"])
 @login_required
 def toggleKey(request):
-    id=request.GET["id"]
+    id=request.POST.get("id")
+    if not id:
+        return HttpResponse("Error: You are missing a key", status=403)
     q=UserPasskey.objects.filter(user=request.user, id=id)
     if q.count()==1:
         key=q[0]
